Add PDF report templates for various entities
- Created `pdf.blade.php` for appointment reports with structured tables and styling.
- Added `receta.blade.php` for medical prescriptions, including patient and medication details.
- Implemented `pdf.blade.php` for doctor reports, listing doctors with their details.
- Developed `pdf.blade.php` for patient discharge reports, including patient and doctor information.
- Created `pdf.blade.php` for patient admission reports, detailing admission and status.
- Added `pdf.blade.php` for patient reports, listing patient details and documents.
- Implemented `pdf.blade.php` for payment reports, detailing appointments and payment statuses.
- Created `recibo.blade.php` for payment receipts, including payment details and status.
- Developed `pdf.blade.php` for role and permission reports, listing roles with associated permissions.
- Added `pdf.blade.php` for medical follow-up reports, detailing patient follow-ups and statuses.
- Created `pdf.blade.php` for specialty reports, listing specialties with their statuses.
- Implemented `pdf.blade.php` for staff reports, detailing staff information without IDs.

+++ app/Http/Controllers/Admin/Doctor/DoctorsController.php
@@ -9,6 +9,7 @@ use App\Models\Doctor\Specialitie;
 use Illuminate\Support\Facades\DB;
 use Spatie\Permission\Models\Role;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Redis;
 use App\Models\Appointment\Appointment;
 use Illuminate\Support\Facades\Storage;
 use App\Models\Doctor\DoctorScheduleDay;
@@ -17,7 +18,8 @@ use App\Models\Doctor\DoctorScheduleHour;
 use App\Http\Resources\User\UserCollection;
 use App\Models\Doctor\DoctorScheduleJoinHour;
 use App\Http\Resources\Appointment\AppointmentCollection;
-use Illuminate\Support\Facades\Redis;
+
+use Barryvdh\DomPDF\Facade\Pdf as PDF;
 
 class DoctorsController extends Controller
 {
@@ -26,6 +28,7 @@ class DoctorsController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAnyDoctor',Doctor::class);
         $search = $request->search;
 
         $users = User::where(DB::raw("CONCAT(users.name,' ',IFNULL(users.surname,''),' ',users.email)"),"like","%".$search."%")
@@ -44,49 +47,7 @@ class DoctorsController extends Controller
     }
 
     public function profile($id){
-        //$user = User::findOrFail($id);
-//
-        //$num_appointment = Appointment::where("doctor_id",$id)->count();
-        //$money_of_appointments = Appointment::where("doctor_id",$id)->sum("amount");
-        //$num_appointment_pendings = Appointment::where("doctor_id",$id)->where("status",1)->count();
-//
-        //$appointment_pendings = Appointment::where("doctor_id",$id)->where("status",1)->get();
-        //$appointments = Appointment::where("doctor_id",$id)->get();
-        //return response()->json([
-        //    "num_appointment" => $num_appointment,
-        //    "money_of_appointments" => $money_of_appointments,
-        //    "num_appointment_pendings" => $num_appointment_pendings,
-        //    "doctor" => UserResource::make($user),
-        //    "appointment_pendings" => AppointmentCollection::make($appointment_pendings),
-        //    "appointments" => $appointments->map(function($appointment){
-        //        return [
-        //            "id" => $appointment->id,
-        //            "patient" => [
-        //                "id" => $appointment->patient->id,
-        //                "full_name" => $appointment->patient->name . ' ' .$appointment->patient->surname,
-        //                "avatar" => $appointment->patient->avatar ? env("APP_URL")."storage/".$appointment->patient->avatar : 'https://cdn-icons-png.flaticon.com/512/1430/1430453.png',
-        //            ],
-        //            "doctor" => [
-        //                "id" => $appointment->doctor->id,
-        //                "full_name" => $appointment->doctor->name . ' ' .$appointment->doctor->surname,
-        //                "avatar" => $appointment->doctor->avatar ? env("APP_URL")."storage/".$appointment->doctor->avatar : NULL,
-        //            ],
-        //            "date_appointment" => $appointment->date_appointment,
-        //            "date_appointment_format" => Carbon::parse($appointment->date_appointment)->format("d M Y"),
-        //            "format_hour_start" => Carbon::parse(date("Y-m-d").' '.$appointment->doctor_schedule_join_hour->doctor_schedule_hour->hour_start)->format("h:i A"),
-        //            "format_hour_end" => Carbon::parse(date("Y-m-d").' '.$appointment->doctor_schedule_join_hour->doctor_schedule_hour->hour_end)->format("h:i A"),
-        //            "appointment_attention" => $appointment->attention ? [
-        //                "id" => $appointment->attention->id,
-        //                "description" => $appointment->attention->description,
-        //                "receta_medica" => $appointment->attention->receta_medica ? json_decode($appointment->attention->receta_medica) : [],
-        //                "created_at" => $appointment->attention->created_at->format("Y-m-d h:i A"),
-        //            ] : NULL,
-        //            "amount" => $appointment->amount,
-        //            "status_pay" => $appointment->status_pay,
-        //            "status" => $appointment->status,
-        //        ];
-        //    }),
-        //]);
+        $this->authorize('profileDoctor',Doctor::class);
         $cachedRecord = Redis::get('profile_doctor_#'.$id);
         $data_doctor = [];
         if(isset($cachedRecord)) {
@@ -178,7 +139,7 @@ class DoctorsController extends Controller
      */
     public function store(Request $request)
     {
-        //dd($request->all());
+        $this->authorize('createDoctor',Doctor::class);
         $schedule_hours = json_decode($request->schedule_hours,1);
         //dd($schedule_hours);
 
@@ -239,6 +200,7 @@ class DoctorsController extends Controller
      */
     public function show(string $id)
     {
+        $this->authorize('viewDoctor',Doctor::class);
         $user = User::findOrFail($id);
 
         return response()->json([
@@ -251,6 +213,7 @@ class DoctorsController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('updateDoctor',Doctor::class);
         $schedule_hours = json_decode($request->schedule_hours,1);
 
         $users_is_valid = User::where("id","<>",$id)->where("email",$request->email)->first();
@@ -425,6 +388,7 @@ class DoctorsController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('deleteDoctor',Doctor::class);
         $user = User::findOrFail($id);
         $user->delete();
         $cachedRecord = Redis::get('profile_doctor_#'.$id);
@@ -434,5 +398,31 @@ class DoctorsController extends Controller
         return response()->json([
             "message" => 200
         ]);
+    }
+
+    public function reporte(Request $request)
+    {
+        $search = $request->search;
+
+        $users = User::where(DB::raw("CONCAT(users.name,' ',IFNULL(users.surname,''),' ',users.email)"),"like","%".$search."%")
+                        // "name","like","%".$search."%"
+                        // ->orWhere("surname","like","%".$search."%")
+                        // ->orWhere("email","like","%".$search."%")
+                        ->orderBy("id","desc")
+                        ->whereHas("roles",function($q){
+                            $q->where("name","like","%DOCTOR%");
+                        })
+                        ->get();
+
+        // dd($users);
+        $resultado = UserCollection::make($users);
+        $resultado = $resultado->toJson();
+        $resultado = json_decode($resultado, true);
+        // dd($resultado);
+        // Generar el PDF
+        // dd($resultado);
+        $pdf = PDF::loadView('ReporteDoctors.pdf', compact('resultado'));
+        // Retornar el PDF como respuesta
+        return $pdf->download('reporte_patients.pdf');
     }
 }
