+++ app/Http/Controllers/Patient/PatientController.php
@@ -14,6 +14,7 @@ use App\Http\Resources\Patient\PatientResource;
 use App\Http\Resources\Patient\PatientCollection;
 use App\Http\Resources\Appointment\AppointmentCollection;
 use Illuminate\Support\Facades\Redis;
+use Barryvdh\DomPDF\Facade\Pdf as PDF;
 
 class PatientController extends Controller
 {
@@ -22,11 +23,12 @@ class PatientController extends Controller
      */
     public function index(Request $request)
     {
+        $this->authorize('viewAny',Patient::class);
         $search = $request->search;
 
-        $patients = Patient::where(DB::raw("CONCAT(patients.name,' ',IFNULL(patients.surname,''),' ',patients.email)"),"like","%".$search."%")
+        $patients = Patient::where(DB::raw("CONCAT(patients.name,' ',IFNULL(patients.surname,''),' ',patients.n_document)"),"like","%".$search."%")
                         ->orderBy("id","desc")
-                        ->paginate(10);
+                        ->paginate(20);
 
         return response()->json([
             "total" => $patients->total(),
@@ -35,50 +37,7 @@ class PatientController extends Controller
     }
 
     public function profile($id) {
-
-        //$patient = Patient::findOrFail($id);
-//
-        //$num_appointment = Appointment::where("patient_id",$id)->count();
-        //$money_of_appointments = Appointment::where("patient_id",$id)->sum("amount");
-        //$num_appointment_pendings = Appointment::where("patient_id",$id)->where("status",1)->count();
-//
-        //$appointment_pendings = Appointment::where("patient_id",$id)->where("status",1)->get();
-        //$appointments = Appointment::where("patient_id",$id)->get();
-        //return response()->json([
-        //    "num_appointment" => $num_appointment,
-        //    "money_of_appointments" => $money_of_appointments,
-        //    "num_appointment_pendings" => $num_appointment_pendings,
-        //    "patient" => PatientResource::make($patient),
-        //    "appointment_pendings" => AppointmentCollection::make($appointment_pendings),
-        //    "appointments" => $appointments->map(function($appointment){
-        //        return [
-        //            "id" => $appointment->id,
-        //            "patient" => [
-        //                "id" => $appointment->patient->id,
-        //                "full_name" => $appointment->patient->name . ' ' .$appointment->patient->surname,
-        //                "avatar" => $appointment->patient->avatar ? env("APP_URL")."storage/".$appointment->patient->avatar : NULL,
-        //            ],
-        //            "doctor" => [
-        //                "id" => $appointment->doctor->id,
-        //                "full_name" => $appointment->doctor->name . ' ' .$appointment->doctor->surname,
-        //                "avatar" => $appointment->doctor->avatar ? env("APP_URL")."storage/".$appointment->doctor->avatar : NULL,
-        //            ],
-        //            "date_appointment" => $appointment->date_appointment,
-        //            "date_appointment_format" => Carbon::parse($appointment->date_appointment)->format("d M Y"),
-        //            "format_hour_start" => Carbon::parse(date("Y-m-d").' '.$appointment->doctor_schedule_join_hour->doctor_schedule_hour->hour_start)->format("h:i A"),
-        //            "format_hour_end" => Carbon::parse(date("Y-m-d").' '.$appointment->doctor_schedule_join_hour->doctor_schedule_hour->hour_end)->format("h:i A"),
-        //            "appointment_attention" => $appointment->attention ? [
-        //                "id" => $appointment->attention->id,
-        //                "description" => $appointment->attention->description,
-        //                "receta_medica" => $appointment->attention->receta_medica ? json_decode($appointment->attention->receta_medica) : [],
-        //                "created_at" => $appointment->attention->created_at->format("Y-m-d h:i A"),
-        //            ] : NULL,
-        //            "amount" => $appointment->amount,
-        //            "status_pay" => $appointment->status_pay,
-        //            "status" => $appointment->status,
-        //        ];
-        //    }),
-        //]);
+        $this->authorize('profile',Patient::class);
         $cachedRecord = Redis::get('profile_patient_#'.$id);
         $data_patient = [];
         if(isset($cachedRecord)) {
@@ -140,6 +99,7 @@ class PatientController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create',Patient::class);
         $patient_is_valid = Patient::where("n_document",$request->n_document)->first();
 
         if($patient_is_valid){
@@ -164,8 +124,22 @@ class PatientController extends Controller
 
         $patient = Patient::create($request->all());
 
-        $request->request->add(["patient_id" => $patient->id]);
-        PatientPerson::create($request->all());
+        // Crear la relación en la tabla 'patient_persons' solo si se proporciona información válida
+        if ($request->has('name_companion')) {
+            $data = $request->only([
+                'name_companion',
+                'surname_companion',
+                'mobile_companion',
+                'relationship_companion',
+                'name_responsible',
+                'surname_responsible',
+                'mobile_responsible',
+                'relationship_responsible',
+            ]);
+            $data['patient_id'] = $patient->id; // Añadir el ID del paciente
+
+            PatientPerson::create($data);
+        }
 
         return response()->json([
             "message" => 200,
@@ -178,6 +152,7 @@ class PatientController extends Controller
      */
     public function show(string $id)
     {
+        $this->authorize('view',Patient::class);
         $patient = Patient::findOrFail($id);
 
         return response()->json([
@@ -190,6 +165,7 @@ class PatientController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('update',Patient::class);
         $patient_is_valid = Patient::where("id","<>",$id)->where("n_document",$request->n_document)->first();
 
         if($patient_is_valid){
@@ -222,9 +198,41 @@ class PatientController extends Controller
         // $request->request->add(["birth_date" => Carbon::parse($request->birth_date, 'GMT')->format("Y-m-d h:i:s")]);
         $patient->update($request->all());
 
-        if($patient->person){
-            $patient->person->update($request->all());
+        // Crear la relación en la tabla 'patient_persons' solo si se proporciona información válida
+        if ($request->has('name_companion')) {
+            $patientPerson = PatientPerson::where('patient_id', $patient->id)->first();
+
+
+
+            if ($patientPerson) {
+                $patientPerson->update($request->only([
+                    'name_companion',
+                    'surname_companion',
+                    'mobile_companion',
+                    'relationship_companion',
+                    'name_responsible',
+                    'surname_responsible',
+                    'mobile_responsible',
+                    'relationship_responsible',
+                ]));
+            }else{
+                // Si el paciente no existe, crea uno nuevo
+                $data = $request->only([
+                    'name_companion',
+                    'surname_companion',
+                    'mobile_companion',
+                    'relationship_companion',
+                    'name_responsible',
+                    'surname_responsible',
+                    'mobile_responsible',
+                    'relationship_responsible',
+                ]);
+                $data['patient_id'] = $patient->id; // Añadir el ID del paciente
+
+                PatientPerson::create($data);
+            }
         }
+
         return response()->json([
             "message" => 200,
             "msg"=>"SE ACTUALIZO EL PACIENTE EXITOSAMENTE"
@@ -236,6 +244,7 @@ class PatientController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete',Patient::class);
         $patient = Patient::findOrFail($id);
         if($patient->avatar){
             Storage::delete($patient->avatar);
@@ -251,5 +260,24 @@ class PatientController extends Controller
             "message" => 200,
             "msg"=>"SE ELIMINO CORRECTAMENTE"
         ]);
+    }
+
+    public function reporte(Request $request)
+    {
+        $this->authorize('viewAny',Patient::class);
+        $search = $request->search;
+
+        $patients = Patient::where(DB::raw("CONCAT(patients.name,' ',IFNULL(patients.surname,''),' ',patients.n_document)"),"like","%".$search."%")
+                        ->orderBy("id","desc")
+                        ->paginate(20);
+
+        $resultado = PatientCollection::make($patients);
+        $resultado = $resultado->toJson();
+        $resultado = json_decode($resultado, true);
+        // Generar el PDF
+        // dd($resultado);
+        $pdf = PDF::loadView('ReportePatient.pdf', compact('resultado'));
+        // Retornar el PDF como respuesta
+        return $pdf->download('reporte_patients.pdf');
     }
 }
